changes from CR
