+++ lbz/rest/__init__.py
@@ -1,2 +1,3 @@
 from lbz.rest.api_gateway_event import APIGatewayEvent
 from lbz.rest.enums import ContentType
+from lbz.rest.request import HTTPRequest
