+++ lbz/websocket/__init__.py
@@ -1,1 +1,2 @@
 from lbz.websocket.enums import ActionType
+from lbz.websocket.request import WebSocketRequest
